Рефакторинг кода, исправляем таймер и по возможности добавляем типы для переменных и функций

+++ lib/views/settings/widgets/internet_settings.dart
@@ -34,7 +34,7 @@ class InternetSettings extends StatelessWidget {
               value: 0,
               groupValue: _settings.internetUsage,
               onChanged: (value) {
-                _settings.internetUsage = value;
+                _settings.internetUsage = value as int;
               },
             ),
           ),
@@ -44,7 +44,7 @@ class InternetSettings extends StatelessWidget {
               value: 2,
               groupValue: _settings.internetUsage,
               onChanged: (value) {
-                _settings.internetUsage = value;
+                _settings.internetUsage = value as int;
               },
             ),
           ),
@@ -54,7 +54,7 @@ class InternetSettings extends StatelessWidget {
               value: 3,
               groupValue: _settings.internetUsage,
               onChanged: (value) {
-                _settings.internetUsage = value;
+                _settings.internetUsage = value as int;
               },
             ),
           ),
